refactor(tests): dissolve tests_review into tests
Merges all tests_review/ files into their natural homes under tests/,
adds session_id fixture to conftest, creates test_router_parsing.py,
and removes the now-empty tests_review/ directory.
