+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-testpaths = tests tests_review
+testpaths = tests
 filterwarnings =
     ignore::DeprecationWarning
