+++ tests/test_costs_flow.py
@@ -5,6 +5,7 @@ import pytest
 from backend.core import db
 from backend.core.models import Tile
 from backend.core.norm_addressees import ADMINISTRATION, BUSINESS, CITIZENS
+from backend.routers.costs import _compute_step_bureaucracy_fractions
 
 
 def test_upsert_process_step_cost_by_addressee_has_no_dead_breakdown_params():
@@ -877,3 +878,170 @@ def test_compute_costs_citizens_ignores_persisted_hourly_rates(test_client):
     assert totals["bureaucracy_cost"] is None
     assert totals["total_time_minutes"] == pytest.approx(60.0)
     assert totals["total_expenses"] == pytest.approx(10.0)
+
+
+# ---------------------------------------------------------------------------
+# session_total_costs_by_addressee: per-addressee independence
+# ---------------------------------------------------------------------------
+
+
+def test_each_addressee_has_independent_totals(session_id):
+    db.upsert_session_total_costs_by_addressee(
+        session_id=session_id,
+        norm_addressee=ADMINISTRATION,
+        total_cost=1234.0,
+        bureaucracy_cost=None,
+        total_time_minutes=None,
+        total_expenses=None,
+    )
+    db.upsert_session_total_costs_by_addressee(
+        session_id=session_id,
+        norm_addressee=BUSINESS,
+        total_cost=5678.0,
+        bureaucracy_cost=900.0,
+        total_time_minutes=None,
+        total_expenses=None,
+    )
+    db.upsert_session_total_costs_by_addressee(
+        session_id=session_id,
+        norm_addressee=CITIZENS,
+        total_cost=None,
+        bureaucracy_cost=None,
+        total_time_minutes=4200.0,
+        total_expenses=150.0,
+    )
+
+    admin = db.get_session_total_costs_by_addressee(session_id, ADMINISTRATION)
+    business = db.get_session_total_costs_by_addressee(session_id, BUSINESS)
+    citizens = db.get_session_total_costs_by_addressee(session_id, CITIZENS)
+
+    assert admin is not None and admin["total_cost"] == 1234.0
+    assert admin["bureaucracy_cost"] is None
+
+    assert business is not None and business["total_cost"] == 5678.0
+    assert business["bureaucracy_cost"] == 900.0
+
+    assert citizens is not None
+    assert citizens["total_cost"] is None, "Buerger duerfen keine Eurokosten haben"
+    assert citizens["total_time_minutes"] == 4200.0
+    assert citizens["total_expenses"] == 150.0
+
+
+def test_upsert_overwrites_only_target_addressee(session_id):
+    db.upsert_session_total_costs_by_addressee(session_id, ADMINISTRATION, 1000.0, None, None, None)
+    db.upsert_session_total_costs_by_addressee(session_id, BUSINESS, 2000.0, 500.0, None, None)
+
+    db.upsert_session_total_costs_by_addressee(session_id, ADMINISTRATION, 9999.0, None, None, None)
+
+    admin = db.get_session_total_costs_by_addressee(session_id, ADMINISTRATION)
+    business = db.get_session_total_costs_by_addressee(session_id, BUSINESS)
+    assert admin["total_cost"] == 9999.0
+    assert business["total_cost"] == 2000.0, "Business-Eintrag darf nicht beruehrt werden"
+
+
+# ---------------------------------------------------------------------------
+# _compute_step_bureaucracy_fractions: proportional allocation formula
+# ---------------------------------------------------------------------------
+
+
+def _seed_business_step_with_regulations(
+    session_id: int,
+    info_count: int,
+    other_count: int,
+) -> int:
+    process_id = db.insert_process(
+        session_id=session_id,
+        process="Wirtschaftsprozess",
+        description="d",
+        norm_addressee=BUSINESS,
+    )
+    case_group_id = db.insert_case_group(
+        session_id=session_id,
+        process_id=process_id,
+        case_group="CG",
+        description="d",
+        norm_addressee=BUSINESS,
+    )
+    step_id = db.insert_process_step(
+        session_id=session_id,
+        case_group_id=case_group_id,
+        step="Schritt",
+        description="d",
+        norm_addressee=BUSINESS,
+    )
+
+    regulation_ids: list[int] = []
+    for i in range(info_count):
+        rid = db.insert_regulation(
+            session_id=session_id,
+            legal_citation=f"§ info-{i}",
+            description=f"Infopflicht {i}",
+            process_id=process_id,
+            applies_to_administration=False,
+            applies_to_business=True,
+            is_business_information_obligation=True,
+        )
+        regulation_ids.append(rid)
+    for i in range(other_count):
+        rid = db.insert_regulation(
+            session_id=session_id,
+            legal_citation=f"§ other-{i}",
+            description=f"Materielle Vorgabe {i}",
+            process_id=process_id,
+            applies_to_administration=False,
+            applies_to_business=True,
+            is_business_information_obligation=False,
+        )
+        regulation_ids.append(rid)
+
+    db.replace_process_step_regulation_links(
+        session_id=session_id,
+        step_id=step_id,
+        norm_addressee=BUSINESS,
+        regulation_ids=regulation_ids,
+    )
+    return step_id
+
+
+def test_pure_information_duty_step_gets_full_fraction(session_id):
+    step_id = _seed_business_step_with_regulations(session_id, info_count=2, other_count=0)
+
+    fractions = _compute_step_bureaucracy_fractions(session_id, BUSINESS)
+
+    assert fractions.get(step_id) == pytest.approx(1.0)
+
+
+def test_mixed_step_gets_proportional_fraction_half(session_id):
+    """A step with one information obligation and one material regulation must
+    receive 0.5 as its fraction, not 0."""
+    step_id = _seed_business_step_with_regulations(session_id, info_count=1, other_count=1)
+
+    fractions = _compute_step_bureaucracy_fractions(session_id, BUSINESS)
+
+    assert fractions.get(step_id) == pytest.approx(0.5)
+
+
+def test_mixed_step_with_one_of_three_info_gets_one_third(session_id):
+    step_id = _seed_business_step_with_regulations(session_id, info_count=1, other_count=2)
+
+    fractions = _compute_step_bureaucracy_fractions(session_id, BUSINESS)
+
+    assert fractions.get(step_id) == pytest.approx(1.0 / 3.0)
+
+
+def test_pure_material_step_is_not_in_fractions(session_id):
+    step_id = _seed_business_step_with_regulations(session_id, info_count=0, other_count=2)
+
+    fractions = _compute_step_bureaucracy_fractions(session_id, BUSINESS)
+
+    assert step_id not in fractions
+
+
+def test_step_without_information_duties_in_session_returns_empty_dict(session_id):
+    """When the session has no information obligations at all, the function
+    returns early with an empty dict."""
+    _seed_business_step_with_regulations(session_id, info_count=0, other_count=3)
+
+    fractions = _compute_step_bureaucracy_fractions(session_id, BUSINESS)
+
+    assert fractions == {}
