+++ tests/test_effort_flow.py
@@ -1158,3 +1158,234 @@ def test_calculate_effort_rejects_invalid_norm_addressee(test_client):
     assert "Unsupported norm_addressee" in resp.json()["detail"]
 
 
+# ---------------------------------------------------------------------------
+# Pay-rate defaults: citizens guard
+# ---------------------------------------------------------------------------
+
+
+def test_get_default_pay_rates_for_citizens_raises():
+    with pytest.raises(ValueError, match="pay-rate defaults"):
+        db.get_default_pay_rates_for_addressee(CITIZENS)
+
+
+def test_get_default_pay_rates_for_unknown_addressee_raises():
+    # normalize_norm_addressee raises first; either ValueError path is acceptable.
+    with pytest.raises(ValueError):
+        db.get_default_pay_rates_for_addressee("unknown_addressee")
+
+
+def test_get_default_pay_rates_for_business_still_works():
+    defaults = db.get_default_pay_rates_for_addressee(BUSINESS)
+    assert set(defaults.keys()) == {"a", "b", "c", "d"}
+    assert all(isinstance(v, float) for v in defaults.values())
+
+
+def test_get_default_pay_rates_for_administration_still_works():
+    defaults = db.get_default_pay_rates_for_addressee(ADMINISTRATION)
+    assert set(defaults.keys()) == {"a", "b", "c", "d"}
+    assert all(isinstance(v, float) for v in defaults.values())
+
+
+# ---------------------------------------------------------------------------
+# Citizens effort upsert: hourly rates must not be persisted
+# ---------------------------------------------------------------------------
+
+
+import itertools
+
+from backend.routers.effort import _resolve_effort_group
+
+
+def _seed_citizens_step(session_id: int) -> int:
+    process_id = db.insert_process(
+        session_id=session_id,
+        process="Buergerprozess",
+        description="d",
+        norm_addressee=CITIZENS,
+    )
+    case_group_id = db.insert_case_group(
+        session_id=session_id,
+        process_id=process_id,
+        case_group="CG",
+        description="d",
+        norm_addressee=CITIZENS,
+    )
+    return db.insert_process_step(
+        session_id=session_id,
+        case_group_id=case_group_id,
+        step="Schritt",
+        description="d",
+        norm_addressee=CITIZENS,
+    )
+
+
+def _empty_rates() -> dict[str, float | None]:
+    return {"a": None, "b": None, "c": None, "d": None}
+
+
+def _empty_times() -> dict[str, float | None]:
+    return {"a": 30.0, "b": None, "c": None, "d": None}
+
+
+def test_citizens_upsert_with_only_time_succeeds(session_id):
+    step_id = _seed_citizens_step(session_id)
+    db.upsert_process_step_effort_split_by_addressee(
+        session_id=session_id,
+        step_id=step_id,
+        norm_addressee=CITIZENS,
+        hourly_rates_current=_empty_rates(),
+        time_required_current=_empty_times(),
+        expenses_current=12.0,
+        hourly_rates_proposed=_empty_rates(),
+        time_required_proposed=_empty_times(),
+        expenses_proposed=12.0,
+    )
+
+
+@pytest.mark.parametrize(
+    "phase,slot",
+    list(itertools.product(["current", "proposed"], ["a", "b", "c", "d"])),
+)
+def test_citizens_upsert_rejects_any_hourly_rate_slot(session_id, phase, slot):
+    step_id = _seed_citizens_step(session_id)
+    rates_current = _empty_rates()
+    rates_proposed = _empty_rates()
+    target = rates_current if phase == "current" else rates_proposed
+    target[slot] = 42.0
+
+    with pytest.raises(ValueError, match="Citizens effort must not persist hourly rates"):
+        db.upsert_process_step_effort_split_by_addressee(
+            session_id=session_id,
+            step_id=step_id,
+            norm_addressee=CITIZENS,
+            hourly_rates_current=rates_current,
+            time_required_current=_empty_times(),
+            expenses_current=None,
+            hourly_rates_proposed=rates_proposed,
+            time_required_proposed=_empty_times(),
+            expenses_proposed=None,
+        )
+
+
+# ---------------------------------------------------------------------------
+# clear_effort_metrics: must only touch target addressee
+# ---------------------------------------------------------------------------
+
+
+def _seed_case_group_with_metrics(session_id: int, norm_addressee: str) -> int:
+    process_id = db.insert_process(
+        session_id=session_id,
+        process="P",
+        description="d",
+        norm_addressee=norm_addressee,
+    )
+    case_group_id = db.insert_case_group(
+        session_id=session_id,
+        process_id=process_id,
+        case_group="CG",
+        description="d",
+        norm_addressee=norm_addressee,
+    )
+    db.upsert_case_group_metrics_by_addressee(
+        session_id=session_id,
+        case_group_id=case_group_id,
+        norm_addressee=norm_addressee,
+        addressees_current=100.0,
+        annual_frequency_current=2.0,
+        cases_current=200.0,
+        addressees_proposed=100.0,
+        annual_frequency_proposed=2.0,
+        cases_proposed=200.0,
+    )
+    return case_group_id
+
+
+@pytest.mark.parametrize("target_addressee", [ADMINISTRATION, BUSINESS, CITIZENS])
+def test_clear_effort_metrics_only_touches_target_addressee(session_id, target_addressee):
+    seeded = {
+        addressee: _seed_case_group_with_metrics(session_id, addressee)
+        for addressee in (ADMINISTRATION, BUSINESS, CITIZENS)
+    }
+
+    db.clear_effort_metrics(session_id, target_addressee)
+
+    for addressee, case_group_id in seeded.items():
+        groups = db.list_case_groups_for_session_and_addressee(session_id, addressee)
+        row = next(g for g in groups if int(g["case_group_id"]) == case_group_id)
+        if addressee == target_addressee:
+            assert row["addressees_current"] is None
+            assert row["annual_frequency_current"] is None
+            assert row["cases_current"] is None
+        else:
+            assert row["addressees_current"] == 100.0
+            assert row["annual_frequency_current"] == 2.0
+            assert row["cases_current"] == 200.0
+
+
+# ---------------------------------------------------------------------------
+# Admin Laufbahn mapping: alias resolution and error handling
+# ---------------------------------------------------------------------------
+
+
+from fastapi import HTTPException as _HTTPException
+
+
+@pytest.mark.parametrize(
+    "raw_value,expected_slot",
+    [
+        ("mittlerer Dienst", "a"),
+        ("m. d.", "a"),
+        ("m.d.", "a"),
+        ("MD", "a"),
+        ("einfacher Dienst", "a"),
+        ("gehobener Dienst", "b"),
+        ("g. d.", "b"),
+        ("GD", "b"),
+        ("hoeherer Dienst", "c"),
+        ("höherer Dienst", "c"),
+        ("h. d.", "c"),
+        ("HD", "c"),
+        ("Durchschnitt", "d"),
+    ],
+)
+def test_admin_laufbahn_aliases_resolve(raw_value, expected_slot):
+    assert _resolve_effort_group({"rolle": raw_value}, ADMINISTRATION) == expected_slot
+
+
+def test_admin_unknown_laufbahn_raises_422():
+    with pytest.raises(_HTTPException) as excinfo:
+        _resolve_effort_group({"rolle": "beliebiger Quark"}, ADMINISTRATION)
+    assert excinfo.value.status_code == 422
+    assert "Unbekannte Laufbahn" in excinfo.value.detail
+
+
+def test_unknown_lohngruppe_letter_raises_422():
+    with pytest.raises(_HTTPException) as excinfo:
+        _resolve_effort_group({"lohngruppe": "e"}, ADMINISTRATION)
+    assert excinfo.value.status_code == 422
+    assert "Unbekannte Lohngruppe" in excinfo.value.detail
+
+
+def test_empty_input_returns_none():
+    assert _resolve_effort_group({}, ADMINISTRATION) is None
+    assert _resolve_effort_group({"rolle": ""}, ADMINISTRATION) is None
+    assert _resolve_effort_group({"rolle": "   "}, ADMINISTRATION) is None
+
+
+def test_valid_lohngruppe_letter_resolves_directly():
+    assert _resolve_effort_group({"lohngruppe": "a"}, ADMINISTRATION) == "a"
+    assert _resolve_effort_group({"lohngruppe": "D"}, ADMINISTRATION) == "d"
+
+
+@pytest.mark.parametrize(
+    "raw_value,expected_slot",
+    [("niedrig", "a"), ("low", "a"), ("mittel", "b"), ("medium", "b"), ("hoch", "c"), ("high", "c")],
+)
+def test_business_level_aliases_resolve(raw_value, expected_slot):
+    assert _resolve_effort_group({"niveau": raw_value}, BUSINESS) == expected_slot
+
+
+def test_business_unknown_level_raises_422():
+    with pytest.raises(_HTTPException) as excinfo:
+        _resolve_effort_group({"niveau": "extrem"}, BUSINESS)
+    assert excinfo.value.status_code == 422
