+++ tests/test_payload_builders.py
@@ -1,3 +1,4 @@
+from backend.core.norm_addressees import ADMINISTRATION, BUSINESS, CITIZENS
 from backend.core.payload_builders import (
     build_case_groups_payload,
     build_processes_payload_with_regulations,
@@ -514,3 +515,76 @@ def test_build_step_analysis_payload_omits_effort_fields_when_present():
     assert "zeitaufwand_in_min_d_vorschlag" not in taetigkeit
     assert "sachaufwand_vorschlag" not in taetigkeit
     assert "ausfuehrung_pro_einzelfall" not in taetigkeit
+
+
+def _multi_na_regulation():
+    return {
+        "regulation_id": 1,
+        "legal_citation": "§1",
+        "description": "Multi-NA",
+        "change_status": "eingefuehrt",
+        "applies_to_administration": 1,
+        "applies_to_business": 1,
+        "applies_to_citizens": 0,
+        "is_business_information_obligation": 0,
+        "process_id": 10,
+    }
+
+
+def test_build_vorgaben_payload_without_filter_keeps_all_addressees():
+    payload = build_vorgaben_payload([_multi_na_regulation()])
+    assert sorted(payload[0]["normadressaten"]) == sorted([ADMINISTRATION, BUSINESS])
+
+
+def test_build_vorgaben_payload_projects_onto_business():
+    payload = build_vorgaben_payload([_multi_na_regulation()], norm_addressee=BUSINESS)
+    assert payload[0]["normadressaten"] == [BUSINESS]
+
+
+def test_build_vorgaben_payload_projects_onto_administration():
+    payload = build_vorgaben_payload(
+        [_multi_na_regulation()], norm_addressee=ADMINISTRATION
+    )
+    assert payload[0]["normadressaten"] == [ADMINISTRATION]
+
+
+def test_build_processes_payload_projects_regulations():
+    processes = [{"process_id": 10, "process": "P", "description": "d"}]
+    payload = build_processes_payload_with_regulations(
+        processes, [_multi_na_regulation()], norm_addressee=BUSINESS
+    )
+    assert payload[0]["vorgaben"][0]["normadressaten"] == [BUSINESS]
+
+
+def test_build_case_groups_payload_projects_regulations():
+    processes = [{"process_id": 10, "process": "P", "description": "d"}]
+    payload = build_case_groups_payload(
+        processes=processes,
+        case_groups=[],
+        regulations=[_multi_na_regulation()],
+        norm_addressee=BUSINESS,
+    )
+    assert payload[0]["vorgaben"][0]["normadressaten"] == [BUSINESS]
+
+
+def test_build_step_analysis_payload_projects_regulations():
+    processes = [{"process_id": 10, "process": "P", "description": "d"}]
+    payload = build_step_analysis_payload(
+        processes=processes,
+        case_groups=[],
+        steps=[],
+        regulations=[_multi_na_regulation()],
+        norm_addressee=BUSINESS,
+    )
+    assert payload[0]["vorgaben"][0]["normadressaten"] == [BUSINESS]
+
+
+def test_projection_fallback_when_no_overlap():
+    # Defensive edge-case: regulation does not belong to the run's NA.
+    # Upstream should have filtered it; builder catches it anyway so an empty
+    # array is never silently sent to the LLM.
+    reg = _multi_na_regulation()
+    reg["applies_to_business"] = 0  # Admin only
+    payload = build_vorgaben_payload([reg], norm_addressee=CITIZENS)
+    # No overlap -> falls back to original list
+    assert payload[0]["normadressaten"] == [ADMINISTRATION]
